Only remove src/default_config.cfg after build if we are in a develop copy
In order to not pollute src/; without this change a script doing on a
sdist-built archive:

    ./setup.py build
    ./setup.py install

would fail because src/default_config.cfg was removed after setup(), and
data/default_config.cfg does not exist

+++ setup.py
@@ -106,6 +106,8 @@ setup(name="poezio",
                       'Screen autoaway plugin': 'pyinotify==0.9.4'})
 
 # Remove the link afterwards
-if exists(join(current_dir, 'src', 'default_config.cfg')):
+if (exists(join(current_dir, 'src', 'default_config.cfg')) and
+        exists(join(current_dir, 'data', 'default_config.cfg'))):
+
     unlink(join(current_dir, 'src', 'default_config.cfg'))
